xds.ins is what I give to shelext xds_a.{ins, res} is the final files
I noticed some things, maybe they will be helpful:

1. If we look at our scripts, we specify the unit cell parameters when we prepare the ins file.
These are the parameters that I have in my data from dials, and they are on the picture I sent.

2. The unit cell parameters of individual datasets found by dials are smaller than parameters that we provided. And they are closer to the parameters from XRD.

3. When we process data with dials, we specify only the space group. In xds it's impossible to specify the space group without the unit cell.
If xds tries to identify the space group automatically, it chooses the wrong one, but it also suggests other space groups with unit cell parameters, so it is possible to choose manually.

4. I tried to process 1 dataset with all the same input parameters, but still got different unit cell.