Testing various ways of processing data with DIALS:
1.  `dials.integrate exclude_images_multiple=N`, then scaling as before
2.  `dials.integrate exclude_images_multiple=N`, then scaling with `intensity_choice=profile`
3.  `dials.slice_sequence exclude_images_multiple=N` after refinement, then integration
4.  `dials.slice_sequence exclude_images_multiple=N`, then further refinement, then integration

I find that 1 and 2 are very similar. 3 and 4 are fragile. The most useful
choice seems to be version 2, which excludes the calibration images from
integration, then explicitly sets dials.scale to use profile fitting
intensities only, as we know removing the calibration images causes many
summation integration intensity estimates to fail.

To avoid the vagaries of shelxt, here don't try to solve the structure, but
refine against a reasonably complete model. I say reasonably complete because
I believe there should be 3 more oxygens and a hydrogen.